requirements.txt
fastapi==0.115.5
torch==2.5.1
torchvision==0.20.1
pillow==10.2.0
uvicorn==0.32.0
python-multipart==0.0.17